+++ Control/Joint/Effects/Either.hs
@@ -2,7 +2,7 @@ module Control.Joint.Effects.Either where
 
 import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
-import Control.Joint.Abilities.Transformer (Transformer (embed, build, unite), Schema, (:>) (T))
+import Control.Joint.Abilities.Transformer (Transformer (build, unite), Schema, (:>) (T))
 import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
 import Control.Joint.Schemes (UT (UT))
 
@@ -13,7 +13,6 @@ instance Interpreted (Either e) where
 type instance Schema (Either e) = UT (Either e)
 
 instance Transformer (Either e) where
-	embed x = T . UT $ Right <$> x
 	build x = T . UT . pure $ x
 	unite = T . UT