+++ Control/Joint/Effects/Maybe.hs
@@ -3,7 +3,7 @@ module Control.Joint.Effects.Maybe where
 import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Completable (Completable (complete))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
-import Control.Joint.Abilities.Transformer (Transformer (embed, build, unite), Schema, (:>) (T))
+import Control.Joint.Abilities.Transformer (Transformer (build, unite), Schema, (:>) (T))
 import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
 import Control.Joint.Schemes (UT (UT))
 
@@ -14,7 +14,6 @@ instance Interpreted Maybe where
 type instance Schema Maybe = UT Maybe
 
 instance Transformer Maybe where
-	embed x = T . UT $ Just <$> x
 	build x = T . UT . pure $ x
 	unite = T . UT