+++ Control/Joint/Effects/State.hs
@@ -5,7 +5,7 @@ import Control.Applicative (Alternative (empty, (<|>)))
 import Control.Joint.Core (type (:.), type (:=))
 import Control.Joint.Abilities.Completable (Completable (complete))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
-import Control.Joint.Abilities.Transformer (Transformer (embed, build, unite), Schema, (:>) (T))
+import Control.Joint.Abilities.Transformer (Transformer (build, unite), Schema, (:>) (T))
 import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
 import Control.Joint.Schemes (TUT (TUT))
 import Control.Joint.Effects.Reader (Reader (Reader))
@@ -35,7 +35,6 @@ instance Interpreted (State s) where
 type instance Schema (State s) = TUT ((->) s) ((,) s)
 
 instance Transformer (State s) where
-	embed x = T . TUT $ \s -> (s,) <$> x
 	build x = T . TUT $ pure <$> run x
 	unite = T . TUT