+++ CHANGELOG.md
@@ -54,3 +54,4 @@
 * Define `MonadTrans` instances for `TU` and `UT` schemes
 * Define `Store` datatype and define `Functor` and `Comonad` instances for it
 * Extract `Schema` associated type family from `Transformer` typeclass
+* Replace `embed` method with `lift` from `MonadTrans` typeclass